requirements.txt
Django==2.1b1
django-db-constraints==0.3.0
paho-mqtt==1.3.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
